+++ embedded-graphics/src/pixelcolor/binary_color.rs
@@ -148,7 +148,7 @@ impl From<bool> for BinaryColor {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::pixelcolor::{Rgb565, RgbColor};
+    use crate::pixelcolor::{IntoStorage, Rgb565, RgbColor};
 
     #[test]
     fn default_color_is_off() {
@@ -198,5 +198,11 @@ mod tests {
 
         assert!(!BinaryColor::Off.is_on());
         assert!(BinaryColor::On.is_on());
+    }
+
+    #[test]
+    fn into_storage() {
+        assert_eq!(BinaryColor::Off.into_storage(), 0u8);
+        assert_eq!(BinaryColor::On.into_storage(), 1u8);
     }
 }