+++ embedded-graphics/CHANGELOG.md
@@ -10,6 +10,17 @@ Embedded Graphics is a `no_std` library for adding graphics features to display 
 
 ### Added
 
+- [#257](https://github.com/jamwaffles/embedded-graphics/pull/257) added the `.into_storage()` method on `PixelColor`s via the `IntoStorage` trait to allow for more ergonomic conversion from a color to its raw storage type.
+
+  ```diff
+  // IntoStorage is included in the prelude.
+  // You can also import it with use embedded_graphics::pixelcolor::IntoStorage.
+  + use embedded_graphics::prelude::*;
+
+  - RawU1::from(color).into_inner()
+  + color.into_storage()
+  ```
+
 - Added as many `#[derive()]`s as possible to all embedded-graphics, tinybmp and tinytga types.
 - Added `From<Point> for [i32; 2]`
 - Added `From<Size> for [u32; 2]`
@@ -410,6 +421,6 @@ A big release, focussed on ergonomics. There are new macros to make drawing and 
   ```
 
 <!-- next-url -->
-[unreleased]: https://github.com/jamwaffles/embedded-graphics/compare/embedded-graphics-v0.6.0-beta.1...HEAD
 
+[unreleased]: https://github.com/jamwaffles/embedded-graphics/compare/embedded-graphics-v0.6.0-beta.1...HEAD
 [0.6.0-beta.1]: https://github.com/jamwaffles/embedded-graphics/compare/embedded-graphics-v0.6.0-alpha.3...embedded-graphics-v0.6.0-beta.1